add logout to check if the updated is correct
